backend/Foodgram/api/urls.py
from rest_framework import routers
from django.urls import include, path

from api.views import UserViewSet, TagViewSet, RecipeViewSet, IngredientViewSet

router = routers.DefaultRouter()
router.register('users', UserViewSet)
router.register('tags', TagViewSet)
router.register('ingredients', IngredientViewSet)
router.register('recipes', RecipeViewSet)


urlpatterns = [
    path('auth/', include('djoser.urls.authtoken')),
    path('', include(router.urls)),
]
