backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.2.0
Pillow==9.0.0
drf-extra-fields==3.6.1
gunicorn==20.1.0
django-filter==21.1